requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
djoser>=2.2.0
django-filter>=23.3
django-cors-headers>=4.2.0
Pillow>=10.0.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0